fix: fMP4 VisualSampleEntry pre_defined field size (2 bytes, not 4)
CMAFWriterCodecConfig wrote writeInt32(-1) for the pre_defined field
which is defined as int(16) in ISO 14496-12 §12.1.3. The extra 2 bytes
shifted the avcC box, causing ffplay/ffprobe to report is_avc=false.

Added BinaryWriter.writeInt16(_:), 2 init segment validation tests.
HEVC (hvc1) was already correct. Version bumped to 0.5.0.

+++ Sources/HLSKit/LiveSegmenter/CMAFWriterCodecConfig.swift
@@ -159,8 +159,8 @@ extension CMAFWriter {
         entry.writeZeros(32)
         // Depth (24 bit)
         entry.writeUInt16(0x0018)
-        // Pre-defined
-        entry.writeInt32(-1)
+        // Pre-defined = -1 (ISO 14496-12 §12.1.3: int(16))
+        entry.writeInt16(-1)
         // avcC box
         entry.writeData(buildAvcC(config: config))
         var box = BinaryWriter()
